:lipstick: Footer now sticks to bottom
also added some padding around contents
