+++ src/components/layout.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import { Box } from '@chakra-ui/core';
 
 import Header from './header';
 import Footer from './footer';
-import Footer2 from './footer2';
 
 interface Props {
     children: React.ReactNode;
@@ -11,11 +11,13 @@ interface Props {
 
 const Layout = ({ children }: Props): JSX.Element => {
     return (
-        <>
+        <Box pos="relative" minHeight="100vh">
             <Header />
-            {children}
-            <Footer2 />
-        </>
+            <Box pl="10%" pr="10%" pb={['450px', '260px']}>
+                {children}
+            </Box>
+            <Footer />
+        </Box>
     );
 };
