+++ src/gatsby-plugin-chakra-ui/theme.js
@@ -33,6 +33,18 @@ const palette = {
         600: '#003C42',
         700: '#001A19',
     },
+    mardiGras: {
+        50: '#E6BFE5',
+        100: '#CE8FCE',
+        200: '#B560B8',
+        300: '#9D30A1',
+        400: '#84008B',
+        500: '#76007D',
+        600: '#67006F',
+        700: '#590061',
+        800: '#4B0053',
+        900: '#3C0045',
+    },
 };
 
 // global config
